Move To drupal cores the steps to check user roles.
