+++ src/Behat/Cores/Drupal7.php
@@ -8,6 +8,7 @@ use Metadrop\Behat\Cores\Traits\CronTrait;
 
 class Drupal7 extends OriginalDrupal7 implements CoreInterface {
 
+  use UsersTrait;
   use CronTrait;
   /**
    * {@inheritdoc}
@@ -36,4 +37,27 @@ class Drupal7 extends OriginalDrupal7 implements CoreInterface {
   public function runElysiaCronJob($job) {
     elysia_cron_run_job($job, TRUE, TRUE, TRUE);
   }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function loadUserByProperty($property, $value, $reset = TRUE) {
+    $query = db_select('users');
+    $query->fields('users', array('uid'));
+    $query->condition($property, $value);
+
+    $result = $query->execute();
+    $uid    = $result->fetchField();
+
+    $account = user_load($uid, $reset);
+    return $account;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getUserRoles($user) {
+    return $user->roles;
+  }
+
 }
